For EXP-2012: Move tests for lib into their own test files
